+++ euporie/__main__.py
@@ -13,20 +13,24 @@ def main() -> "None":
 
     App: "Type[EuporieApp]"
 
-    is_dumb_term = os.environ.get("TERM") == "dumb"
-    if is_dumb_term and not config.dump:
-        config.defaults["page"] = True
-    if is_dumb_term:
-        config.defaults["dump"] = True
-    if not config.dump and not sys.stdin.isatty() and not sys.stderr.isatty():
-        config.defaults["dump"] = True
-
     # If on a dumb terminal, the TUI app won't work, so dump and notebooks instead
+    if not config.dump:
+        is_dumb_term = os.environ.get("TERM") == "dumb"
+        have_tty = sys.stdin.isatty() or sys.stderr.isatty()
+        # If interaction is not supported but dumping was not asked for, dump and page
+        if is_dumb_term or not have_tty:
+            config.defaults["dump"] = True
+            config.defaults["page"] = True
+
     if config.dump:
 
         from euporie.app.dump import DumpApp
 
         App = DumpApp
+    elif config.hub:
+        from euporie.app.hub import HubApp
+
+        App = HubApp
     else:
         from euporie.app.tui import TuiApp