+++ euporie/app/tui.py
@@ -12,7 +12,6 @@ from prompt_toolkit.clipboard.pyperclip import PyperclipClipboard
 from prompt_toolkit.completion import PathCompleter
 from prompt_toolkit.filters import Condition, buffer_has_focus
 from prompt_toolkit.formatted_text import HTML, fragment_list_to_text, to_formatted_text
-from prompt_toolkit.input.defaults import create_input
 from prompt_toolkit.key_binding.key_bindings import KeyBindings, merge_key_bindings
 from prompt_toolkit.layout import (
     ConditionalContainer,
@@ -28,7 +27,6 @@ from prompt_toolkit.layout import (
 from prompt_toolkit.layout.controls import FormattedTextControl
 from prompt_toolkit.layout.dimension import Dimension
 from prompt_toolkit.layout.menus import CompletionsMenu
-from prompt_toolkit.output.defaults import create_output
 from prompt_toolkit.widgets import Button, Dialog, Label, SearchToolbar, TextArea
 from pyperclip import determine_clipboard  # type: ignore
 
@@ -51,10 +49,8 @@ if TYPE_CHECKING:
     from prompt_toolkit.clipboard import Clipboard
     from prompt_toolkit.completion import Completer
     from prompt_toolkit.formatted_text import AnyFormattedText, StyleAndTextTuples
-    from prompt_toolkit.input import Input
     from prompt_toolkit.key_binding.key_processor import KeyPressEvent
     from prompt_toolkit.layout.containers import AnyContainer
-    from prompt_toolkit.output import Output
 
     from euporie.tabs.base import Tab
     from euporie.tabs.notebook import Notebook
@@ -73,10 +69,14 @@ class TuiApp(EuporieApp):
     def __init__(self, **kwargs: "Any") -> "None":
         """Create a new euporie text user interface application instance."""
         super().__init__(
-            full_screen=True,
-            mouse_support=True,
-            editing_mode=self.get_edit_mode(),
-            **kwargs,
+            **{
+                **{
+                    "full_screen": True,
+                    "mouse_support": True,
+                    "editing_mode": self.get_edit_mode(),
+                },
+                **kwargs,
+            }
         )
         self.has_dialog = False
 
@@ -86,24 +86,6 @@ class TuiApp(EuporieApp):
             await asyncio.sleep(config.terminal_polling_interval)
             self.term_info.background_color.send()
             self.term_info.foreground_color.send()
-
-    def load_input(self) -> "Input":
-        """Creates the input for this application to use.
-
-        Returns:
-            A prompt-toolkit input instance
-
-        """
-        return create_input(always_prefer_tty=True)
-
-    def load_output(self) -> "Output":
-        """Creates the output for this application to use.
-
-        Returns:
-            A prompt-toolkit output instance
-
-        """
-        return create_output(always_prefer_tty=True)
 
     def load_clipboard(self) -> "None":
         """Determines which clipboard mechanism to use."""
@@ -232,7 +214,8 @@ class TuiApp(EuporieApp):
             ],
         )
         is_searching = Condition(
-            lambda: self.search_bar.control in self.layout.search_links
+            lambda: self.search_bar is not None
+            and self.search_bar.control in self.layout.search_links
         )
 
         status_bar = ConditionalContainer(